Move `RemoteFileSyncer` and `RemoteFileProcessor` to separate package.
