+++ src/main/java/org/prebid/server/execution/file/RemoteFileProcessor.java
@@ -1,4 +1,4 @@
-package org.prebid.server.execution;
+package org.prebid.server.execution.file;
 
 import io.vertx.core.Future;
