+++ src/main/java/org/prebid/server/execution/file/RemoteFileSyncer.java
@@ -1,4 +1,4 @@
-package org.prebid.server.execution;
+package org.prebid.server.execution.file;
 
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.vertx.core.Future;
