+++ src/main/java/org/prebid/server/geolocation/MaxMindGeoLocationService.java
@@ -14,7 +14,7 @@ import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 import org.apache.commons.compress.archivers.tar.TarArchiveInputStream;
 import org.apache.commons.lang3.StringUtils;
-import org.prebid.server.execution.RemoteFileProcessor;
+import org.prebid.server.execution.file.RemoteFileProcessor;
 import org.prebid.server.execution.timeout.Timeout;
 import org.prebid.server.geolocation.model.GeoInfo;
